requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
numpy==1.26.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
